Challenge_1a/requirements.txt
pymupdf==1.23.7
lingua-language-detector==2.0.2
sentence-transformers==2.5.1
torch==2.1.2
scikit-learn==1.3.0
numpy==1.24.3